Restructure source code into package and modules
