+++ Pipfile
@@ -8,6 +8,7 @@ opencv-python = "*"
 matplotlib = "*"
 numpy = "*"
 statsmodels = "*"
+pandas = "*"
 
 [dev-packages]
