--- features.txt
@@ -1,101 +1,0 @@
-upper_zone,middle_zone,lower_zone,median_width,most_occurring_angle,average_angle,stdev_angle
-0.3305084745762712,0.2796610169491525,0.3898305084745763,22.5,49.5,48.9,3.9
-0.3793103448275862,0.3017241379310345,0.31896551724137934,16.5,49.5,45.7,7.1
-0.3018867924528302,0.3584905660377358,0.33962264150943394,16,40.5,44.6,8.2
-0.40336134453781514,0.2857142857142857,0.31092436974789917,18.0,49.5,47.2,5.3
-0.30973451327433627,0.3274336283185841,0.36283185840707965,21.5,49.5,49.5,4.3
-0.13186813186813187,0.4175824175824176,0.45054945054945056,21.5,49.5,53.5,4.3
-0.3779527559055118,0.2992125984251969,0.3228346456692913,16.5,49.5,51.8,3.3
-0.10843373493975904,0.43373493975903615,0.4578313253012048,18,54.0,52.3,3.6
-0.31851851851851853,0.31851851851851853,0.362962962962963,24,54.0,59.6,6.8
-0.24299065420560748,0.3925233644859813,0.3644859813084112,25.0,40.5,52.9,11.8
-0.3900709219858156,0.3049645390070922,0.3049645390070922,25.0,49.5,60.3,17.1
-0.38016528925619836,0.3140495867768595,0.30578512396694213,18,18.0,43.7,22.2
-0.248,0.36,0.392,21.5,49.5,53.5,6.5
-0.19491525423728814,0.3728813559322034,0.4322033898305085,19.0,49.5,45.4,11.7
-0.34782608695652173,0.35507246376811596,0.2971014492753623,21.5,63.0,52.2,11.3
-0.24817518248175183,0.35766423357664234,0.39416058394160586,21.5,63.0,60.0,3.7
-0.27358490566037735,0.32075471698113206,0.4056603773584906,18.5,18.0,43.9,19.1
-0.3037974683544304,0.2721518987341772,0.4240506329113924,23.5,175.5,0,0
-0.21621621621621623,0.3063063063063063,0.4774774774774775,12.5,72.0,59.6,9.5
-0.40476190476190477,0.32142857142857145,0.27380952380952384,16.5,63.0,56.4,10.0
-0.36666666666666664,0.3,0.3333333333333333,14.5,58.5,55.8,6.0
-0.2714285714285714,0.37142857142857144,0.35714285714285715,18.0,58.5,54.7,15.0
-0.4766355140186916,0.2616822429906542,0.2616822429906542,15.0,58.5,55.7,5.3
-0.425,0.3375,0.2375,13,58.5,50.5,12.3
-0.345679012345679,0.3333333333333333,0.32098765432098764,15,63.0,59.1,5.7
-0.46153846153846156,0.2857142857142857,0.25274725274725274,16.0,58.5,53.4,12.6
-0.4574468085106383,0.2978723404255319,0.24468085106382978,16,63.0,60.7,2.4
-0.4731182795698925,0.3010752688172043,0.22580645161290322,14,63.0,55.9,11.1
-0.5476190476190477,0.25,0.20238095238095238,21,90.0,91.5,2.6
-0.4896551724137931,0.30344827586206896,0.20689655172413793,20,90.0,90.0,0.0
-0.40404040404040403,0.3838383838383838,0.21212121212121213,18.0,90.0,118.0,45.5
-0.5142857142857142,0.2785714285714286,0.20714285714285716,16.0,90.0,91.5,2.3
-0.5823529411764706,0.25882352941176473,0.1588235294117647,19.0,94.5,93.0,2.6
-0.39814814814814814,0.37962962962962965,0.2222222222222222,18.5,90.0,69.2,29.7
-0.4225352112676056,0.34507042253521125,0.2323943661971831,18,94.5,94.5,0.0
-0.45222929936305734,0.2929936305732484,0.25477707006369427,17,94.5,96.3,2.5
-0.5068493150684932,0.3219178082191781,0.17123287671232876,20.5,175.5,158.4,38.2
-0.5193370165745856,0.24861878453038674,0.23204419889502761,17.5,175.5,0,0
-0.23958333333333334,0.3645833333333333,0.3958333333333333,12.5,58.5,61.2,5.3
-0.23863636363636365,0.29545454545454547,0.4659090909090909,12,49.5,56.2,12.8
-0.20430107526881722,0.4838709677419355,0.3118279569892473,14.5,58.5,66.0,16.2
-0.33636363636363636,0.32727272727272727,0.33636363636363636,12,58.5,57.8,6.0
-0.34782608695652173,0.34782608695652173,0.30434782608695654,11,72.0,55.8,12.6
-0.3434343434343434,0.29292929292929293,0.36363636363636365,12.5,63.0,58.5,14.5
-0.29896907216494845,0.32989690721649484,0.3711340206185567,12,54.0,60.4,9.7
-0.2857142857142857,0.22321428571428573,0.49107142857142855,8.5,72.0,69.2,9.0
-0.3068181818181818,0.3068181818181818,0.38636363636363635,11.0,58.5,64.0,19.2
-0.38181818181818183,0.3,0.3181818181818182,14,49.5,50.6,2.2
-0.34615384615384615,0.33653846153846156,0.3173076923076923,19.0,58.5,60.0,9.3
-0.26804123711340205,0.28865979381443296,0.44329896907216493,11,72.0,59.6,10.0
-0.40186915887850466,0.2897196261682243,0.308411214953271,10.5,58.5,55.5,5.2
-0.27835051546391754,0.29896907216494845,0.422680412371134,13,63.0,63.0,12.6
-0.32673267326732675,0.25742574257425743,0.4158415841584158,8.0,54.0,63.0,10.1
-0.25806451612903225,0.25806451612903225,0.4838709677419355,12.0,63.0,65.0,15.7
-0.32075471698113206,0.330188679245283,0.3490566037735849,14,67.5,55.7,11.0
-0.125,0.3,0.575,12.0,49.5,54.8,5.0
-0.3425925925925926,0.24074074074074073,0.4166666666666667,12.5,63.0,60.2,3.3
-0.3076923076923077,0.2403846153846154,0.4519230769230769,11.0,54.0,62.0,13.1
-0.37383177570093457,0.2336448598130841,0.3925233644859813,9.5,58.5,64.8,8.0
-0.32673267326732675,0.31683168316831684,0.3564356435643564,10.5,76.5,63.6,10.0
-0.3,0.35555555555555557,0.34444444444444444,11,94.5,72.0,22.9
-0.26881720430107525,0.4731182795698925,0.25806451612903225,11.5,72.0,63.0,11.7
-0.31896551724137934,0.3275862068965517,0.35344827586206895,15,76.5,66.6,12.1
-0.32,0.38,0.3,11.5,4.5,34.1,27.4
-0.2767857142857143,0.33035714285714285,0.39285714285714285,10.5,36.0,46.1,17.4
-0.2661290322580645,0.3709677419354839,0.3629032258064516,12.5,175.5,0,0
-0.15217391304347827,0.43478260869565216,0.41304347826086957,12,54.0,59.5,8.1
-0.3333333333333333,0.25833333333333336,0.4083333333333333,9,54.0,54.0,0
-0.32407407407407407,0.28703703703703703,0.3888888888888889,13.0,67.5,64.5,3.7
-0.2736842105263158,0.2736842105263158,0.45263157894736844,14.0,63.0,67.5,9.9
-0.10526315789473684,0.3815789473684211,0.5131578947368421,13,49.5,53.2,4.9
-0.14606741573033707,0.3258426966292135,0.5280898876404494,12,49.5,57.3,9.4
-0.2978723404255319,0.30851063829787234,0.39361702127659576,10.5,67.5,60.7,9.2
-0.32673267326732675,0.37623762376237624,0.297029702970297,15.0,54.0,54.3,9.5
-0.2571428571428571,0.3333333333333333,0.4095238095238095,11.0,58.5,60.7,9.8
-0.24489795918367346,0.3163265306122449,0.4387755102040816,14,67.5,64.8,6.8
-0.32,0.25,0.43,8.0,63.0,56.3,7.8
-0.2403846153846154,0.3557692307692308,0.40384615384615385,11.0,58.5,59.0,9.4
-0.33653846153846156,0.3173076923076923,0.34615384615384615,13,4.5,24.6,31.1
-0.39705882352941174,0.29411764705882354,0.3088235294117647,10.5,58.5,60.0,2.6
-0.3277310924369748,0.29411764705882354,0.37815126050420167,11.5,9.0,24.4,23.8
-0.4032258064516129,0.2903225806451613,0.3064516129032258,10.5,72.0,62.1,14.4
-0.3983739837398374,0.2601626016260163,0.34146341463414637,12.0,49.5,47.6,9.3
-0.4444444444444444,0.2564102564102564,0.29914529914529914,9,67.5,57.2,13.1
-0.3103448275862069,0.28448275862068967,0.4051724137931034,12,36.0,47.1,13.3
-0.34782608695652173,0.2782608695652174,0.3739130434782609,11,54.0,54.0,10.4
-0.4198473282442748,0.26717557251908397,0.31297709923664124,13.0,63.0,60.7,3.2
-0.3442622950819672,0.2786885245901639,0.3770491803278688,12.5,58.5,61.5,5.5
-0.37815126050420167,0.2689075630252101,0.35294117647058826,10.0,63.0,73.6,14.7
-0.46017699115044247,0.30973451327433627,0.23008849557522124,13,67.5,54.8,16.3
-0.45081967213114754,0.27049180327868855,0.2786885245901639,11.0,54.0,47.2,9.5
-0.34579439252336447,0.32710280373831774,0.32710280373831774,11.5,63.0,60.0,16.6
-0.43089430894308944,0.2926829268292683,0.2764227642276423,13.0,58.5,50.2,12.3
-0.3619047619047619,0.3238095238095238,0.3142857142857143,13.5,36.0,48.5,10.9
-0.18823529411764706,0.4,0.4117647058823529,11.5,40.5,55.4,13.0
-0.1267605633802817,0.5211267605633803,0.352112676056338,9,54.0,71.0,19.5
-0.4174757281553398,0.30097087378640774,0.2815533980582524,21,112.5,112.2,4.5
-0.3983050847457627,0.2627118644067797,0.3389830508474576,19.5,112.5,114.4,8.8
-0.43119266055045874,0.28440366972477066,0.28440366972477066,14.0,112.5,112.1,4.2
-0.4107142857142857,0.25892857142857145,0.33035714285714285,16.5,117.0,114.5,3.1
